dev fix (Теперь сортировка не сбивается при навигации)

+++ Services/Admins/DeleteAdminService.php
@@ -9,7 +9,7 @@ class DeleteAdminService extends BaseService
 {
     public function delete(int $id): void
     {
-        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+        if ($_SERVER['REQUEST_METHOD'] === 'POST' && $_SESSION['user']['role'] !== '0') {
             $warning = 'The article was`nt deleted!' . "\n";
 
             if ($this->repository->delete($id)) {
@@ -21,6 +21,7 @@ class DeleteAdminService extends BaseService
             } else {
                 $_SESSION['warning'] = $warning;
             }
+
             \header("Location: {$_SERVER['HTTP_REFERER']}");
         } else {
             \header('Location: /');
